@@ -1,53 +1,29 @@
 {
-  "name": "colab-editor",
-  "version": "0.1.0",
-  "private": true,
-  "dependencies": {
-    "@testing-library/jest-dom": "^4.2.4",
-    "@testing-library/react": "^9.3.2",
-    "@testing-library/user-event": "^7.1.2",
-    "prosemirror-commands": "^1.1.4",
-    "prosemirror-dropcursor": "^1.3.2",
-    "prosemirror-gapcursor": "^1.1.5",
-    "prosemirror-history": "^1.1.3",
-    "prosemirror-inputrules": "^1.1.2",
-    "prosemirror-keymap": "^1.1.4",
-    "prosemirror-menu": "^1.1.4",
-    "prosemirror-model": "^1.11.0",
-    "prosemirror-schema-list": "^1.1.2",
-    "prosemirror-state": "^1.3.3",
-    "prosemirror-view": "^1.15.2",
-    "react": "^16.13.1",
-    "react-dom": "^16.13.1",
-    "react-intl": "^2.9.0",
-    "react-scripts": "3.4.1",
-    "rxjs-compat": "^6.6.0",
-    "ws": "^7.3.0",
-    "y-prosemirror": "^0.3.3",
-    "y-protocols": "^0.2.3",
-    "y-websocket": "^1.3.1",
-    "yjs": "^13.0.0"
-  },
+  "name": "collab-editor-client",
+  "version": "1.0.0",
+  "description": "Collaborative editor client with HTML formatted text editing",
   "scripts": {
-    "start": "set NODE_OPTIONS=--openssl-legacy-provider && react-scripts start",
-    "server": "nodemon --exec babel-node server",
-    "build": "set NODE_OPTIONS=--openssl-legacy-provider && react-scripts build",
-    "test": "react-scripts test",
-    "eject": "react-scripts eject"
+    "start": "webpack serve --mode development --open",
+    "build": "webpack --mode production"
   },
-  "eslintConfig": {
-    "extends": "react-app"
+  "dependencies": {
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "react-bootstrap": "^2.8.0",
+    "bootstrap": "^5.3.1",
+    "yjs": "^13.5.0",
+    "y-websocket": "^1.4.5"
   },
-  "browserslist": {
-    "production": [
-      ">0.2%",
-      "not dead",
-      "not op_mini all"
-    ],
-    "development": [
-      "last 1 chrome version",
-      "last 1 firefox version",
-      "last 1 safari version"
-    ]
+  "devDependencies": {
+    "@babel/core": "^7.22.10",
+    "@babel/preset-env": "^7.22.10",
+    "@babel/preset-react": "^7.22.5",
+    "babel-loader": "^9.1.3",
+    "css-loader": "^6.8.1",
+    "style-loader": "^3.3.3",
+    "html-webpack-plugin": "^5.5.3",
+    "webpack": "^5.88.2",
+    "webpack-cli": "^5.1.4",
+    "webpack-dev-server": "^4.15.1"
   }
 }
